Migrating operator_test to ote
